refactoring notes controllers
