+++ index.js
@@ -6,15 +6,13 @@ const Tracing = require('@sentry/tracing')
 const express = require('express')
 const app = express()
 const cors = require('cors')
-const User = require('./models/User')
-const Note = require('./models/Note')
 
 const notFound = require('./middleware/notFound.js')
 const handleErrors = require('./middleware/handleErrors.js')
-const userExtractor = require('./middleware/userExtractor')
 
 const usersRouter = require('./controllers/users')
 const loginRouter = require('./controllers/login')
+const notesRouter = require('./controllers/notes')
 
 app.use(cors())
 app.use(express.json())
@@ -47,92 +45,7 @@ app.get('/', (request, response) => {
   response.send('<h1>Hello World!</h1>')
 })
 
-app.get('/api/notes', async (request, response) => {
-  const notes = await Note.find({}).populate('user', {
-    username: 1,
-    name: 1
-  })
-  response.json(notes)
-})
-
-app.get('/api/notes/:id', (request, response, next) => {
-  const { id } = request.params
-
-  Note.findById(id)
-    .then(note => {
-      if (note) return response.json(note)
-      response.status(404).end()
-    })
-    .catch(err => next(err))
-})
-
-app.put('/api/notes/:id', userExtractor, (request, response, next) => {
-  const { id } = request.params
-  const note = request.body
-
-  const newNoteInfo = {
-    content: note.content,
-    important: note.important
-  }
-
-  Note.findByIdAndUpdate(id, newNoteInfo, { new: true })
-    .then(result => {
-      response.json(result)
-    })
-    .catch(next)
-})
-
-app.delete('/api/notes/:id', userExtractor, async (request, response, next) => {
-  const { id } = request.params
-  // const note = await Note.findById(id)
-  // if (!note) return response.sendStatus(404)
-
-  const res = await Note.findByIdAndDelete(id)
-  if (res === null) return response.sendStatus(404)
-
-  response.status(204).end()
-})
-
-app.post('/api/notes', userExtractor, async (request, response, next) => {
-  const {
-    content,
-    important = false
-  } = request.body
-
-  // sacar userId de request
-  const { userId } = request
-
-  const user = await User.findById(userId)
-
-  if (!content) {
-    return response.status(400).json({
-      error: 'required "content" field is missing'
-    })
-  }
-
-  const newNote = new Note({
-    content,
-    date: new Date(),
-    important,
-    user: user._id
-  })
-
-  // newNote.save().then(savedNote => {
-  //   response.json(savedNote)
-  // }).catch(err => next(err))
-
-  try {
-    const savedNote = await newNote.save()
-
-    user.notes = user.notes.concat(savedNote._id)
-    await user.save()
-
-    response.json(savedNote)
-  } catch (error) {
-    next(error)
-  }
-})
-
+app.use('/api/notes', notesRouter)
 app.use('/api/users', usersRouter)
 app.use('/api/login', loginRouter)
